requirements.txt
streamlit
numpy
pandas
Pillow
gspread
oauth2client